Fixed* add backround noise
as it turns out when you would use it inside the custom audiomentations compose class the augmentation didnt do anything. Also moved the test
