+++ epochalyst/pipeline/model/training/augmentation/utils.py
@@ -9,8 +9,6 @@ Classes:
 from dataclasses import dataclass, field
 from typing import Any
 
-import numpy as np
-import numpy.typing as npt
 import torch
 
 from epochalyst.pipeline.model.training.utils.recursive_repr import recursive_repr
@@ -131,34 +129,6 @@ class NoOp(torch.nn.Module):
         -------
             torch.Tensor: The augmented input signal tensor.
         """
-        return x
-
-
-@dataclass
-class AddBackgroundNoiseWrapper:
-    """Wrapper class to be used for audiomentations AddBackgroundNoise augmentation."""
-
-    p: float = 0.5
-    sounds_path: str = field(default="data/raw/", repr=False)
-    dataset_name: str = ""
-    min_snr_db: float = -3.0
-    max_snr_db: float = 3.0
-    aug: Any = None
-
-    def __post_init__(self) -> None:
-        """Post initialization function of AddBackgroundNoiseWrapper."""
-        self.aug = get_audiomentations().AddBackgroundNoise(
-            p=self.p,
-            sounds_path=self.sounds_path,
-            min_snr_db=self.max_snr_db,
-            max_snr_db=self.max_snr_db,
-            noise_transform=get_audiomentations().PolarityInversion(p=0.5),
-        )
-
-    def __call__(self, x: npt.NDArray[np.float32], sr: int) -> npt.NDArray[np.float32]:
-        """Apply the augmentation to the input signal."""
-        if self.aug is not None:
-            return self.aug(x, sr)
         return x
