+++ tests/pipeline/model/training/augmentation/test_audio/white_noise.wav
